refactors contract-related logic into a separate package
